requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Web Scraping
beautifulsoup4==4.12.2

# Data Models
pydantic==2.5.0
pydantic-settings==2.1.0

# Search API
tavily-python==0.3.0

# File Processing
pyyaml==6.0.1

# Document Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0

# OCR Processing (Optional)
pytesseract==0.3.10
Pillow==10.1.0

# Environment Variables
python-dotenv==1.0.0

# Vector Database
weaviate-client==3.25.3

# Async utilities
aiofiles==23.2.1

# Date utilities
python-dateutil==2.8.2

# OpenTelemetry (Observability)
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-otlp==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-httpx==0.42b0

# Development Dependencies (for CI/CD)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
ruff==0.1.6
mypy==1.7.1
bandit==1.7.5
safety==3.2.8
